upgrade some thing

+++ onlineBookStore/src/main/java/com/TechYash_Bit/onlineBookStore/Repositories/BookRepo.java
@@ -6,7 +6,7 @@ import org.springframework.data.jpa.repository.Query;
 
 import java.util.List;
 
-public interface BookRepo extends JpaRepository<BookEntity,Integer> {
+public interface BookRepo extends JpaRepository<BookEntity,Long> {
     //@Query("select b.title as title,b.author as author,b.price as price from BookEntity b")
     //List<TitleAutorPriceDto> getBook();
 }
