+++ onlineBookStore/src/main/java/com/TechYash_Bit/onlineBookStore/Repositories/OrderRepo.java
@@ -3,5 +3,5 @@ package com.TechYash_Bit.onlineBookStore.Repositories;
 import com.TechYash_Bit.onlineBookStore.Entities.OrderEntity;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface OrderRepo extends JpaRepository<OrderEntity,Integer> {
+public interface OrderRepo extends JpaRepository<OrderEntity,Long> {
 }
